+++ app/Http/Controllers/Blog/Post/DeleteController.php
@@ -11,9 +11,16 @@ class DeleteController extends Controller
     public function delete($slug){
 
         $post = Post::fetchBySlug($slug);
+        if($post->publisher->user_id == auth()->user()->id){
+
         $post = $post->delete();
 
         return redirect(route('blog.show'));
+        }
+        
+        else{
+            return redirect()->back();
+        }
 
     }
 }
